feat: Add options for formatting control
This commit introduces new options to the formatter to allow for greater
control over the output:

- `InlineArrays`: Formats arrays on a single line. - `InlineStrings`:
Formats strings on a single line. - `UseFieldCommas`: Adds commas to the
end of fields in objects and arrays. - `UseTrailingCommas`: Adds a
trailing comma to the end of the last field in objects and arrays, if
`UseFieldCommas` is also enabled.